Extract confirmation toggle into hui-action-confirmation-toggle component
Moves the inline toggle + pencil row out of hui-action-editor into a
dedicated hui-action-confirmation-toggle component in confirmation-editor/.
